requirements.txt
asn1crypto==1.0.0
backports.csv==1.0.7
certifi==2019.11.28
cffi==1.12.3
chardet==3.0.4
cryptography==2.7
defusedxml==0.6.0
dicttoxml==1.7.4
diff-match-patch==20181111
Django==2.2.5
django-import-export==1.2.0
django-random-queryset==0.1.2
django-rest-knox==4.1.0
djangorestframework==3.10.3
djangorestframework-simplejwt==4.3.0
et-xmlfile==1.0.1
geoip2==3.0.0
httplib2==0.17.0
idna==2.8
jdcal==1.4.1
joblib==0.14.1
maxminddb==1.5.2
numpy==1.18.3
odfpy==1.4.0
openpyxl==3.0.0
pandas==1.0.3
Pillow==6.1.0
pycparser==2.19
Pygments==2.4.2
PyJWT==1.7.1
python-dateutil==2.8.1
pytils==0.3
pytz==2019.2
PyYAML==5.1.2
requests==2.22.0
scikit-learn==0.22.2.post1
scipy==1.4.1
six==1.12.0
sklearn==0.0
sqlparse==0.3.0
tablib==0.13.0
urllib3==1.25.7
xlrd==1.2.0
xlwt==1.3.0
